+++ yandex/data_source_yandex_vpc_security_group_rule_test.go
@@ -1,0 +1,58 @@
+package yandex
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
+	"testing"
+)
+
+func TestAccDataSourceVPCSecurityGroupRule(t *testing.T) {
+	t.Parallel()
+
+	sgr1Name := "data.yandex_vpc_security_group_rule.rule1"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckVPCSecurityGroupDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccDataSourceVPCSecurityGroupRuleResourceConfig(),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(sgr1Name, "description", "rule1 description"),
+					resource.TestCheckResourceAttr(sgr1Name, "direction", "ingress"),
+					resource.TestCheckResourceAttr(sgr1Name, "port", "8080"),
+					resource.TestCheckResourceAttr(sgr1Name, "protocol", "TCP"),
+					resource.TestCheckResourceAttr(sgr1Name, "v4_cidr_blocks.#", "2"),
+					resource.TestCheckResourceAttr(sgr1Name, "v4_cidr_blocks.0", "10.0.1.0/24"),
+					resource.TestCheckResourceAttr(sgr1Name, "v4_cidr_blocks.1", "10.0.2.0/24"),
+				),
+			},
+		},
+	})
+}
+
+func testAccDataSourceVPCSecurityGroupRuleResourceConfig() string {
+	return `
+resource "yandex_vpc_network" "net" {}
+
+resource "yandex_vpc_security_group" "sg" {
+  network_id  = "${yandex_vpc_network.net.id}"
+  name        = "some name"
+  description = "some description"
+}
+
+resource "yandex_vpc_security_group_rule" "rule" {
+  security_group_binding = "${yandex_vpc_security_group.sg.id}"
+  description    = "rule1 description"
+  direction      = "ingress"
+  protocol       = "TCP"
+  v4_cidr_blocks = ["10.0.1.0/24", "10.0.2.0/24"]
+  port           = 8080
+}
+
+data "yandex_vpc_security_group_rule" "rule1" {
+  security_group_binding = "${yandex_vpc_security_group.sg.id}"
+  rule_id                = "${yandex_vpc_security_group_rule.rule.id}"
+}
+`
+}